fix lmconfigure update
Change parameters passed to update endpoint
to be of type Optional[str] = Body(None)
where before we used Body(). Also includes
changes to check if an argument is not None,
and changed the endpoint to not split list
args, because they are split for us in the
lmconfigure update command.